Implemented new functionality for the stage creator: now it can change and place amongst all wall types, not just rocks; it can select enemy mode and place, delete, move and select enemies; can also swap between all types of enemies; can also place, delete, move and edit screen warps, and it can also add screen warp destinations in the new exclusive screen warp menu "screen warp configurator". Centralized enemies and walls lists in enums lists.

+++ core/src/main/java/com/mygdx/game/items/TextureManager.java
@@ -290,11 +290,11 @@ public class TextureManager {
 		videos.clear();
 	}
 
-	public static void fixatedText (String text,float x, float y,int timeOnScreen,int size){
+	public static void fixatedText (String text,float x, float y,int timeOnScreen,float size){
 		TextureManager.fixatedText.add(new Text(text,x,y,size,timeOnScreen));
 	}
 
-	public static void fixatedText (String text,float x, float y,int timeOnScreen,int size,int r, int g, int b){
+	public static void fixatedText (String text,float x, float y,int timeOnScreen,float size,int r, int g, int b){
 		TextureManager.fixatedText.add(new Text(text,x,y, timeOnScreen,r,g,b,1,0,size));
 	}
